remove dulplicate style code, put Contract first display
