+++ src/pages/wallet/wallet.connect/modal/contract.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import {
-  View, Text, StyleSheet, TouchableOpacity, Linking,
+  View, Text, TouchableOpacity, Linking,
 } from 'react-native';
 import PropTypes from 'prop-types';
 import Rsk3 from '@rsksmart/rsk3';
@@ -9,42 +9,12 @@ import _ from 'lodash';
 
 import { strings } from '../../../../common/i18n';
 import BaseModal from './base';
-import color from '../../../../assets/styles/color';
-import fontFamily from '../../../../assets/styles/font.family';
 import { WALLET_CONNECT } from '../../../../common/constants';
 import config from '../../../../../config';
 import common from '../../../../common/common';
+import styles from '../../../../assets/styles/dapp.popup';
 
 const { MODAL_TYPE } = WALLET_CONNECT;
-
-const styles = StyleSheet.create({
-  line: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 16,
-  },
-  lineTitle: {
-    color: color.black,
-    fontSize: 15,
-    fontFamily: fontFamily.AvenirBook,
-  },
-  lineValue: {
-    color: color.dustyGray,
-    fontSize: 15,
-    fontFamily: fontFamily.AvenirBook,
-    width: '60%',
-    textAlign: 'right',
-  },
-  toAddressLink: {
-    width: '60%',
-    alignSelf: 'flex-end',
-  },
-  addressLineValue: {
-    width: '100%',
-    color: color.app.theme,
-    fontSize: 13,
-  },
-});
 
 export default class Contract extends Component {
   onToAddressPressed = (toAddress) => {
@@ -120,20 +90,20 @@ export default class Contract extends Component {
               ) : (
                 <>
                   {
-                    abiInputData.method !== null && (
-                      <View style={styles.line}>
-                        <Text style={styles.lineTitle}>{strings('page.wallet.walletconnect.method')}</Text>
-                        <Text style={styles.lineValue}>{abiInputData.method}</Text>
-                      </View>
-                    )
-                  }
-                  {
                     _.isEmpty(abiInputData.params.Contract) && (
                       <View style={styles.line}>
                         <Text style={styles.lineTitle}>{strings('page.wallet.walletconnect.contract')}</Text>
                         <TouchableOpacity style={styles.toAddressLink} onPress={() => this.onToAddressPressed(to)}>
                           <Text style={[styles.lineValue, styles.addressLineValue]}>{to}</Text>
                         </TouchableOpacity>
+                      </View>
+                    )
+                  }
+                  {
+                    abiInputData.method !== null && (
+                      <View style={styles.line}>
+                        <Text style={styles.lineTitle}>{strings('page.wallet.walletconnect.method')}</Text>
+                        <Text style={styles.lineValue}>{abiInputData.method}</Text>
                       </View>
                     )
                   }
